requirements.txt
gspread==6.2.1
google-auth-oauthlib==1.2.3
google-auth==2.41.1
pandas==2.3.3
gspread-dataframe==4.0.0